third_party/WebKit/LayoutTests/fast/events/touch/gesture/long-press-focuses-frame-expected.txt

CONSOLE WARNING: Resource requests whose URLs contain raw newline characters are deprecated, and may be blocked in M60, around August 2017. Please remove newlines from places like element attribute values in order to continue loading those resources. See https://www.chromestatus.com/features/5735596811091968 for more details.
PASS successfullyParsed is true

TEST COMPLETE
PASS document.activeElement == frame1 is true
PASS document.activeElement == frame2 is true
PASS document.activeElement == document.body is true
This test checks that long press to select text focuses the pressed frame.
 
